Added 'autotest' mode to healthd service
This mode makes healthd service to configure itself as manager for
a handful of IEEE specializations, and shows agent connections,
measurements etc. in console. This way, simple end-to-end tests
can be carried out without the need of a healthd D-Bus client
(e.g. for some platform in which there is no Python and there
is no capable client already developed).