Added connection/disconnection events to manager listener
Added connection and disconnection events to manager listener.
(Agent listener already had an event for connection. Added one
for disconnection as well.)

This moves a common pattern across plug-ins (notifying the top-level
application about connections and disconnections) to the core,
removing code duplication and the need of special plug-in init.

Added connection/disconnection event fix to Android plug-in