Added BlueZ rediscovering script to sdk
BlueZ caches the list of services offered by a Bluetooth device
upon pairing. If the device is another development computer, it
might happen that HDP service is not up at that moment. In this
case, HDP API won't work for that device until it is paired again
or services are again discovered.

Apps like sample_bt_agent should incorporate a call to
DiscoverServices() in the future to avoid this problem
altogether, and mitigating the need for this script.