Fixed support to BlueZ HDP secondary/streaming channels
Secondary HDP channels are normally streaming (non-reliable) channels
for real-time data where latency is more important than reliability.
This type of channel receives unconfirmed report APDUs and should be
handled as part of the main channel (through which the association is
set up and shot down).

11073 core is notified about connection and disconnection of the main
channel only; secondary channels can open and close at will, and
received APDUs go to the same communication context of the first HDP
channel.

Data is sent via the main channel only (assuming that streaming
channels only bring unconfirmed events and/or confirmations may go
through the main channel; if this is wrong, it needs to be fixed in
the future.