Fix test_float test
Test was assuming that 0.5 has exact representation as floats, but
the read_float() mechanism uses pow(10, -1) * 5 to decode the value
which may result in a small epsilon. The test worked under Valgrind
(which simulates another architecture) but did not work while
running natively. This patch fixes the test for native case.