Fix bugs related to multiple channel handling (HDP, BlueZ)
1) When a disconnection is forced by us, close all channels
related to the device; there may be secondary channels, or
even lingering ones (see note).

2) When sending APDU fails, force a disconnection.

It was observed then, when agent is also BlueZ, deletion of an
old channel was not always properly reported by the remote party.

We refrain from destroying channels (keeping in mind that they might
reconnect, as allowed by MCAP) but this depends on perfect remote
device collaboration (i.e. notification of channel deletion).

When this does not happen, we end up with an old channel besides
the new one, and then our stack tries to send data through the old
one, because it is the first.

In order to fix this, we now force a disconnection in all channels
when an APDU cannot be sent. This cleans up all lingering channels
and makes sure that remote device will start over with a fresh
channel.