Filter out double channel creation
When channel is initiated by us, we get feedback twice: via connection
callback and via signal, and we end up with two file descriptors for
the same channel. In some enviroments it seems harmless, but people
have reported misbehavior related to this (perhaps some race condition?).
In any case, acquiring FD twice is not ok, so we mark the channel as
initiated by us and ignore the signal.

This assumes that ChannelConnected signal always comes after
CreateChannel completion callback (which is the proven current
case; but if it changes in the future, there will be trouble).